[WIP][TASK] Fields need hinting
The core offers validation methods, but there is no possibility to show recommendations for fields.

This patch adds a field for the browser title which uses hinting. This involves:
Validation by hinting rules
Showing the hint on focus
Adding classes to tabs and fields when a field has a hint