[TASK] Red warning color should not be used in hinting
When a hinting rule fails, it will show the hinting in red. Red means danger, which is not the case with hinting. Only green and yellow should be used as signal colors.

This patch replaces the label class where needed and changes the character count range validation not to use the threshold anymore.